Read host processes from the mounted location on /hostfs.
Up until now, Metricbeat wasn't aware of host processes because when we run Metricbeat inside the container it starts without the "-system.hostfs=/hostfs" flag option and no matter what config we set in the system module, it cannot see host processes. The solution is to add this flag as the command inside the docker compose.